# contracts/openapi.yaml
# 생성물 — 수기로 수정하지 말 것. 정본은 apps/api의 Pydantic 모델이다 (ADR-0009 계약 절).
# 재생성: cd apps/api && uv run python scripts/export_openapi.py
openapi: 3.1.0
info:
  title: cookmark-api
  version: 0.1.0
paths:
  /api/v1/health:
    get:
      summary: Health
      operationId: health_api_v1_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Health Api V1 Health Get
  /api/v1/auth/{provider}/login:
    get:
      tags:
      - auth
      summary: Start Login
      description: IdP 인가 화면으로 보낸다. redirect_uri는 라우트에서 도출해 설정 없이 성립시킨다.
      operationId: start_login_api_v1_auth__provider__login_get
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Provider'
      responses:
        '302':
          description: IdP 인가 화면으로 리다이렉트
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/{provider}/callback:
    get:
      tags:
      - auth
      summary: Auth Callback
      description: IdP가 돌려보낸 코드를 검증된 신원으로 바꾸고 세션을 발급한다.
      operationId: auth_callback_api_v1_auth__provider__callback_get
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          description: 세션이 없거나 유효하지 않다
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Read Current Account
      description: 세션 검증 표면이자 로그인 데모 지점.
      operationId: read_current_account_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          description: 세션이 없거나 유효하지 않다
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      description: 제시된 증표를 파기한다. 이미 죽은 증표로 불러도 결과는 같으므로 멱등이다.
      operationId: logout_api_v1_auth_logout_post
      responses:
        '204':
          description: Successful Response
  /api/v1/auth/account:
    delete:
      tags:
      - auth
      summary: Withdraw
      description: 탈퇴 — 계정·세션 즉시 하드 삭제. 지울 대상은 세션에서 나오지 클라이언트가 대지 않는다.
      operationId: withdraw_api_v1_auth_account_delete
      responses:
        '204':
          description: Successful Response
        '401':
          description: 세션이 없거나 유효하지 않다
  /api/v1/llm/recognize:
    post:
      tags:
      - llm
      summary: Recognize
      description: '냉장고 사진에서 재료 후보를 얻는다. 이미지는 메모리로만 지나간다 — 저장·로깅 없음(스펙 #96).'
      operationId: recognize_api_v1_llm_recognize_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecognizeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognizeResponse'
        '400':
          description: 본문 파싱 실패
        '401':
          description: 세션이 없거나 유효하지 않다
        '502':
          description: LLM 업스트림 호출 실패
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/llm/extract:
    post:
      tags:
      - llm
      summary: Extract
      description: 레시피 제목에서 재료를 추론한다 — 제목만 본다(본문·자막 금지).
      operationId: extract_api_v1_llm_extract_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractResponse'
        '400':
          description: 본문 파싱 실패
        '401':
          description: 세션이 없거나 유효하지 않다
        '502':
          description: LLM 업스트림 호출 실패
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/llm/match:
    post:
      tags:
      - llm
      summary: Match
      description: 확정 재료 + 저장 레시피로 후보를 받고 매치 점수를 실산출해 낸다(#101 이월 해소).
      operationId: match_api_v1_llm_match_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
        '400':
          description: 본문 파싱 실패
        '401':
          description: 세션이 없거나 유효하지 않다
        '502':
          description: LLM 업스트림 호출 실패
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/og-image:
    get:
      tags:
      - og-image
      summary: Read Og Image
      description: 출처 페이지의 og:image URL — 제안 카드 음식 사진용. 부재는 명시적 null이다.
      operationId: read_og_image_api_v1_og_image_get
      parameters:
      - name: url
        in: query
        required: true
        schema:
          type: string
          format: uri
          minLength: 1
          maxLength: 2083
          description: 레시피 출처 페이지 URL
          title: Url
        description: 레시피 출처 페이지 URL
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OgImageResponse'
        '401':
          description: 세션이 없거나 유효하지 않다
        '400':
          description: 허용되지 않는 대상 주소
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/recipes:
    get:
      tags:
      - recipes
      summary: List Recipes
      description: 소유자 항목만, 삽입순(created_at·id) — 스코프는 Repository 생성 시점에 이미 끝났다.
      operationId: list_recipes_api_v1_recipes_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RecipeResponse'
                type: array
                title: Response List Recipes Api V1 Recipes Get
        '401':
          description: 세션이 없거나 유효하지 않다
    post:
      tags:
      - recipes
      summary: Create Recipe
      description: 저장하면 제목에서 재료 추출이 1회 일어나 항목에 남는다. 추출 실패는 명시적 502 — 조용한 저장 없음.
      operationId: create_recipe_api_v1_recipes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: 본문을 파싱할 수 없다(JSON 디코드 실패)
        '401':
          description: 세션이 없거나 유효하지 않다
        '502':
          description: 재료 추출에 실패했다 — 레시피는 저장되지 않았다
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/recipes/{recipe_id}:
    get:
      tags:
      - recipes
      summary: Read Recipe
      operationId: read_recipe_api_v1_recipes__recipe_id__get
      parameters:
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Recipe Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '401':
          description: 세션이 없거나 유효하지 않다
        '404':
          description: 레시피를 찾을 수 없다
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - recipes
      summary: Update Recipe
      description: url은 불변이고 재추출은 없다 — 바꿀 수 있는 건 title·ingredients뿐이다.
      operationId: update_recipe_api_v1_recipes__recipe_id__patch
      parameters:
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Recipe Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: 본문을 파싱할 수 없다(JSON 디코드 실패)
        '401':
          description: 세션이 없거나 유효하지 않다
        '404':
          description: 레시피를 찾을 수 없다
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - recipes
      summary: Delete Recipe
      operationId: delete_recipe_api_v1_recipes__recipe_id__delete
      parameters:
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Recipe Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 세션이 없거나 유효하지 않다
        '404':
          description: 레시피를 찾을 수 없다
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AccountResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        iss:
          type: string
          title: Iss
        sub:
          type: string
          title: Sub
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - iss
      - sub
      - created_at
      title: AccountResponse
      description: 서버가 계정에 대해 아는 전부. 더 있는 것처럼 보이지 않게 이게 그대로 계약이다.
    ExtractRequest:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
      type: object
      required:
      - title
      title: ExtractRequest
      description: '추출 요청 — 제목만 받는다(본문·자막은 수익화·법무 리서치 #5로 범위 제한).'
    ExtractResponse:
      properties:
        ingredients:
          items:
            type: string
          type: array
          title: Ingredients
        usage:
          $ref: '#/components/schemas/LLMUsage'
      type: object
      required:
      - ingredients
      - usage
      title: ExtractResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LLMUsage:
      properties:
        prompt_tokens:
          type: integer
          title: Prompt Tokens
        output_tokens:
          type: integer
          title: Output Tokens
        thought_tokens:
          type: integer
          title: Thought Tokens
        image_tokens:
          type: integer
          title: Image Tokens
        cost_usd:
          type: number
          title: Cost Usd
        model:
          type: string
          title: Model
      type: object
      required:
      - prompt_tokens
      - output_tokens
      - thought_tokens
      - image_tokens
      - cost_usd
      - model
      title: LLMUsage
      description: '호출 1건의 사용량 — 필드 구성은 T1 #6 실측 resolution. 토큰을 하나로 뭉치지 않는다.'
    MatchRecipe:
      properties:
        title:
          type: string
          title: Title
        ingredients:
          items:
            type: string
          type: array
          title: Ingredients
      type: object
      required:
      - title
      title: MatchRecipe
      description: 매칭 입력의 저장 레시피 — 제목 + (있으면) 재료 목록. URL은 클라이언트에 남는다.
    MatchRequest:
      properties:
        ingredients:
          items:
            type: string
          type: array
          minItems: 1
          title: Ingredients
        recipes:
          items:
            $ref: '#/components/schemas/MatchRecipe'
          type: array
          title: Recipes
      type: object
      required:
      - ingredients
      title: MatchRequest
      description: 매칭 요청 — ingredients는 체크·비뭉뚱그림만 온다(ADR-0002 필터는 클라이언트 소관).
    MatchResponse:
      properties:
        suggestions:
          items:
            $ref: '#/components/schemas/Suggestion'
          type: array
          title: Suggestions
        usage:
          $ref: '#/components/schemas/LLMUsage'
      type: object
      required:
      - suggestions
      - usage
      title: MatchResponse
      description: 매칭 결과 — 후보 최대 6개를 그대로 낸다. 3개 상한·라벨 결정은 클라이언트 소관(무변경).
    MissingIngredient:
      properties:
        name:
          type: string
          title: Name
        substitute:
          anyOf:
          - type: string
          - type: 'null'
          title: Substitute
      type: object
      required:
      - name
      title: MissingIngredient
    OgImageResponse:
      properties:
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
      type: object
      required:
      - image_url
      title: OgImageResponse
    Provider:
      type: string
      enum:
      - kakao
      - google
      title: Provider
    RecipeCreate:
      properties:
        url:
          type: string
          minLength: 1
          title: Url
        title:
          type: string
          minLength: 1
          title: Title
      additionalProperties: false
      type: object
      required:
      - url
      - title
      title: RecipeCreate
      description: ingredients는 받지 않는다 — 추출은 서버가 저장 시 1회 수행한다. 몰래 넣으면 422로 시끄럽게 거절.
    RecipeResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        url:
          type: string
          title: Url
        title:
          type: string
          title: Title
        ingredients:
          items:
            type: string
          type: array
          title: Ingredients
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - url
      - title
      - ingredients
      - created_at
      title: RecipeResponse
    RecipeUpdate:
      properties:
        title:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Title
        ingredients:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Ingredients
      additionalProperties: false
      type: object
      title: RecipeUpdate
      description: 'url은 불변 — 스키마에 없고 extra=forbid라 보내면 422다(조용한 무시가 아니라 계약). 재추출도
        없다.


        null과 부재를 같게 본다(둘 다 무변경) — ingredients=[]는 null이 아니라 "비우기"다.'
    RecognizeRequest:
      properties:
        image_base64:
          type: string
          format: base64
          contentMediaType: application/octet-stream
          title: Image Base64
      type: object
      required:
      - image_base64
      title: RecognizeRequest
      description: 인식 요청 — 클라이언트가 768px로 리사이즈한 JPEG의 base64. 디코드 실패·빈 이미지는 422다.
    RecognizeResponse:
      properties:
        ingredients:
          items:
            $ref: '#/components/schemas/RecognizedIngredient'
          type: array
          title: Ingredients
        low_quality:
          type: boolean
          title: Low Quality
          default: false
        usage:
          $ref: '#/components/schemas/LLMUsage'
      type: object
      required:
      - ingredients
      - usage
      title: RecognizeResponse
      description: 인식 결과 — seam 반환 타입을 겸한다(후처리가 없어 wire 모델이 곧 경계 타입이다).
    RecognizedIngredient:
      properties:
        name:
          type: string
          title: Name
        confidence:
          type: string
          enum:
          - high
          - medium
          - low
          title: Confidence
      type: object
      required:
      - name
      - confidence
      title: RecognizedIngredient
    SessionResponse:
      properties:
        token:
          type: string
          title: Token
        expires_at:
          type: string
          format: date-time
          title: Expires At
        account:
          $ref: '#/components/schemas/AccountResponse'
      type: object
      required:
      - token
      - expires_at
      - account
      title: SessionResponse
      description: 발급된 세션. 웹은 쿠키로 받으므로 token을 무시하고, 네이티브는 이 값을 보안 저장소에 넣는다 (§9).
    Suggestion:
      properties:
        menu:
          type: string
          title: Menu
        source:
          type: string
          enum:
          - saved
          - generated
          title: Source
        missing:
          items:
            $ref: '#/components/schemas/MissingIngredient'
          type: array
          title: Missing
        reason:
          type: string
          title: Reason
        match_score:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Match Score
      type: object
      required:
      - menu
      - source
      - missing
      - reason
      - match_score
      title: Suggestion
      description: 제안 1건 — match_score는 서버 산식 실산출이다(#101 신규, ADR-0007 이월 해소).
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError